re-org code structure
